requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
django-allauth==0.45.0
mysqlclient==2.0.3
line-bot-sdk==2.4.2